[refactor] Restructure newly introduces types
